requirements.txt
click==8.1.7
packageurl-python @ git+https://github.com/relizaio/packageurl-python@ab5a18320247d848ce5907669b4752f1d20d80c7